committt

+++ user-service/src/main/java/com/m1kellz/userservice/repository/TokenRepository.java
@@ -3,15 +3,7 @@ package com.m1kellz.userservice.repository;
 import com.m1kellz.userservice.entity.Token;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
-
-//
-//  @Query(value = """
-//      select t from Token t inner join User u\s
-//      on t.user.id = u.id\s
-//      where u.id = :id and (t.expired = false or t.revoked = false)\s
-//      """)
-//  List<Token> findAllValidTokenByUser(Integer id);
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
   Token findByToken(String token);
 }
